+++ common/src/main/java/io/github/xienaoban/biologydictionary/BiologyDictionaryClient.java
@@ -31,6 +31,7 @@ public final class BiologyDictionaryClient {
 
     private static final Queue<Component> pendingTextBoxLogs = new ConcurrentLinkedQueue<>();
     private static boolean demoMode = false;
+    private static boolean debugMode = false;
 
     private int ticks;
 
@@ -149,5 +150,12 @@ public final class BiologyDictionaryClient {
     public static boolean toggleDemoMode() {
         demoMode = !demoMode;
         return demoMode;
+    }
+
+    public static boolean isDebugMode() { return debugMode; }
+
+    public static boolean toggleDebugMode() {
+        debugMode = !debugMode;
+        return debugMode;
     }
 }